requirements.txt
openai==1.51.0
pymupdf==1.24.9
pyyaml==6.0.2
tqdm==4.66.5
streamlit==1.37.1
fhir.resources==7.1.0
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.8.2
typing_extensions==4.12.2
httpx==0.27.2
anyio==4.4.0
sniffio==1.3.1
certifi==2024.7.4
numpy==1.26.4
pandas==2.2.2
fastapi==0.115.0
uvicorn==0.30.6